Fix legion_commander; Fix abaddon cast certain abilities on creep heroes; Fix drow_ranger, clinkz, silencer auto-cast abilities cast whenever possible; Prevent crystal_maiden from casting frostbite on antimage_counterspell; Add doom_bringer auto-cast devour when devouring better creeps; Fix jakiro doesn't cast liquid_fire, liquid_frost on buildings; Decrease desire of lich casting frost_armour on creeps; Fix abyssal_underlord casting pit_of_malice on non-teleporting enemeis; Fix mirana casting moon_shadow on creep heroes; Add necrolyte casting reapers_scythe on teleporting enemies; Prevent slardar from casting corrosive_haze on affected units; Decrease desire of undying casting decay on creeps; Change brewmaster, viper ability level-up up to 7.31c; Prevent nyx_assassin casting impale on magic immune targets; Prevent venomancer casting venomous_gale on magic immune targets; Decrease desire of warlock casting upheavel on creeps; Fix typos in winter_wyvern source; Change item build lists of some heroes;

+++ ability_item_usage_abyssal_underlord.lua
@@ -8,6 +8,7 @@
 local utility = require( GetScriptDirectory().."/utility" ) 
 require(GetScriptDirectory() ..  "/ability_item_usage_generic")
 local AbilityExtensions = require(GetScriptDirectory().."/util/AbilityAbstraction")
+local A = require(GetScriptDirectory().."/util/MiraDota")
 
 local debugmode=false
 local npcBot = GetBot()
@@ -28,16 +29,16 @@ local AbilityToLevelUp=
 	Abilities[2],
 	Abilities[1],
 	Abilities[2],
-	Abilities[5],
+	Abilities[4],
 	"talent",
 	Abilities[2],
-	Abilities[5],
+	Abilities[4],
 	Abilities[3],
 	Abilities[3],
 	"talent",
 	Abilities[3],
 	"nil",
-	Abilities[5],
+	Abilities[4],
 	"nil",
 	"talent",
 	"nil",
@@ -247,12 +248,9 @@ Consider[2]=function() --Location AOE Example
 	-- Global high-priorty usage
 	--------------------------------------
 	-- Check for a channeling enemy
-	for _,npcEnemy in pairs( enemys )
-	do
-		if ( npcEnemy:IsChanneling() ) 
-		then
-			return BOT_ACTION_DESIRE_HIGH, npcEnemy:GetLocation();
-		end
+	local channelingEnemy = A.Linq.First(enemys, function(t) A.Hero.IsTeleporting(t) and CanCast[2](t) end)
+	if channelingEnemy then
+		return BOT_ACTION_DESIRE_HIGH, npcEnemy:GetLocation()
 	end
 
 	--try to kill enemy hero
@@ -334,9 +332,9 @@ local darkRiftOriginalTarget
 local trackOriginalTargetPosition
 local darkRiftChosenTarget
 
-Consider[5]=function()
-	
-	local abilityNumber=5
+-- 7.31 rework fiends_gate
+Consider[4]=function()
+	local abilityNumber=4
 	--------------------------------------
 	-- Generic Variable Setting
 	--------------------------------------
@@ -380,7 +378,7 @@ Consider[5]=function()
 				local target = FindNearbyTeleportTarget(npcEnemy)
 				if target ~= nil then
 					darkRiftOriginalTarget = npcEnemy
-					return BOT_ACTION_DESIRE_HIGH, target, "Target"
+					return BOT_ACTION_DESIRE_HIGH, target:GetLocation(), "Location"
 				end
 			end
 		end
@@ -390,8 +388,8 @@ Consider[5]=function()
 end
 
 local darkRiftCancelRange = 1200^2
-Consider[6] = function()
-	local ability = AbilitiesReal[6]
+Consider[5] = function()
+	local ability = AbilitiesReal[5]
 	if not ability:IsFullyCastable() or ability:IsHidden() then
 		return 0
 	end
@@ -411,6 +409,7 @@ Consider[6] = function()
 	end 
 	return 0
 end
+Consider[5] = A.Dota.EmptyDesireFun
 
 AbilityExtensions:AutoModifyConsiderFunction(npcBot, Consider, AbilitiesReal)
 function AbilityUsageThink()
